chore: extract RecordingManager, PopupManager and simplify event dispatch
Continue CamelMonitor God class decomposition (Steps 2, 3, 6):

- Extract RecordingManager (~280 lines): screenshot capture, tape recording,
  keystroke display, screen buffer tracking, key label utility
- Extract PopupManager (~430 lines): switch-integration, more-tabs, and
  kill-confirm popup state management, key handling, and rendering
- Simplify event dispatch: move overview-specific key handlers (p, x, X, r,
  d, f) into OverviewTab via OverviewActions callback interface, removing
  duplicate handlers from CamelMonitor.handleTabKeys
- Update McpFacade to use RecordingManager instead of direct bridge callbacks

CamelMonitor reduced from ~2,593 to ~2,167 lines (16% reduction).
All 260 tests pass. Pure refactor, zero behavior change.
